feat(viz): Update Source Code using the config form
Currently, there's no mechanism for updating the Source Code
through the Design page.

This commit brings that functionality by leveraging the `path`
property from the `IVisualizationNode` object and lodash.set.

The new content is shaped by `uniforms`, using the composed
JSON Schema out of the Camel Components properties.

This is the first iteration, there's more work to be done, as
the serialization process depends on the actual entity.

relates to: https://github.com/KaotoIO/kaoto-next/issues/137