chore(ui): Remove direct catalog usage
Currently, we're using the catalog at runtime by:

```
import catalogIndex from '@kaoto-next/camel-catalog/index.json?url';
```

While it works, the problem with this approach is that we're tying
the @kaoto-next/camel-catalog dependency at runtime.

The change is to leverage the vite build step in which the catalog
is copied to the `assets` folder.

```
    fetch(`.${DEFAULT_CATALOG_PATH}/index.json`)
```

Relates to: https://github.com/KaotoIO/kaoto-next/issues/131