chore(uniforms-patternfly): Remove library
Currently, `@kaoto-next/uniforms-patternfly` resides as a package of this
monorepo.

While this brings better working ergonomics, this prevents to use automatic
versioning as it will means that with every change, we'll be forced to
make a version bump in `@kaoto-next/ui` and sometimes we might want to
pacage more changes in a single version.

The changes is to remove this package in favor of the extracted one here:
https://github.com/KaotoIO/uniforms-patternfly
which supports automatic version and publishing after each merge to the `main` branch