feat(versioning): Version and publish workflow
The `trigger-release.yml` workflow, execute a lerna-lite publish command
which calculate the packages versions, publish them to the npmjs registry,
pushes tags and lastly create GitHub releases for each package.

This commit also changes the `repository` to point to the root repository
URL and `repositoryDirectory` to the actual folder of each package to avoid
problems while generating the release notes.