chore(viznode): Remove usage of the VisualizationNode class
In some cases, we're using the VisualizationNode class directly to create a new
node as `new VisualizationNode(...)`.

This is not a problem itself, but tying entity classes like `KameletBinding` with
the `VisualizationNode` class is not ok, as it allows to implement functionality
directly on the implementation class that other entities won't leverage.

This commit encompass:

* Remove the usage of `new VisualizationClass(...)`
* Create a new function that returns an instance of `VisualizationNode` but using the
`IVisualizationNode` type instead