chore(format): Enable linting on CI
Currently, lint-staged files is only working locally, through a husky hook.

The issue with this approach is that requires that all contributors have
said hooks installed, otherwise, we might introduce unformatted code.

The change is to run the linter directly in each package on the CI, so this
way, any unformated code will be spotted.