feat(catalog): Minify Camel Catalog
Currently, we're generating the Camel Catalog at build time, in a pretty
fashion.

The issue with this approach is the file size increase.

This pull request minifies those files at the UI build time, but preserves
them in a pretty fashion in the `dist` folder, so it can be analyzed locally for devs.

Fixes: https://github.com/KaotoIO/kaoto-next/issues/131