feat(catalog): Introduce Camel Catalog Service
Currently, we're using the CatalogStore to hold all components.

The issue with this approach is that consumers of the public
components will need to install the store dependency and also
to interact with said store in order to add additional components.

The alternative is to split the `CatalogSchemaLoaderProvider` in two
sections, `CatalogProvider` and `SchemasProvider` respectively. This
way, consumers can provide their own components coming from different
sources, without interacting with any store.