feat(viz): Add Data property to IViznode
Currently, we have a few supporting properties in the IVizNode class,
for instance, `label`, `iconData` and `path`.

Now we need dedicated properties, from the underlying `BaseVisualEntity` that
help us to determine which component are we working with, the issue with that
is that we would need to add optional properties from the different underlying
`BaseVisualEntity`s and that's not ideal.

A solution for this scenario is to provide a generic `data` property that will
contain the common properties, like the ones mentioned before, plus any dedicated
property from the underlying `BaseVisualEntity`.

The advantage of doing this is that we can work with those properties inside the
underlying `BaseVisualEntity` and get autocompletion.