feat(multiplying-architecture): Provide an editor using the Multiplying architecture
* Notes
If the new content is the same as the current one, we don't need to update the Editor,
as it will regenerate the Camel Resource, hence disconnecting the configuration form (if open).

This happens due to the multiplying architecture lifecycle, where the content is set
after saving the file.

The lifecycle is:

1.  User edits the file either adding a new node or modifying an existing one using the form
2.  User saves the file
3.  The Envelope uses the `getContent` callback to retrieve the new content
4.  The Envelope sets the new content using the `setContent` callback

At this point, both the new content and the current content are the same, so the Editor
don't need to be updated.

fix: https://github.com/KaotoIO/kaoto-next/issues/291