feat(canvas): Split Canvas fallback component
Currently, the canvas fallback component that is loaded when a route
cannot be parsed, has a link to the SourceCode editor. The issue with
this approach is that some consumers won't have said source code editor.

This commit provides a default fallback component and moves the custom
fallback component to the DesignPage, which is Source Code editor aware.