fix(viz): Configure nodes that could contain children
* Context
There are some EIPs that could contain children, like `choice`, `when`, `otherwise`, `filter` and others.

At this moment, the configuration panel doesn't work for those EIPs because those properties
that signal where to add the children.

* Changes
As this property is handled by the canvas itself, this pull request filter them out from the configuration form.

* Notes
A more robust solution might be added in the future, as new EIPs are added.

Fixes:
- https://github.com/KaotoIO/kaoto-next/issues/320
- https://github.com/KaotoIO/kaoto-next/issues/321

Relates to:
- https://github.com/KaotoIO/kaoto-next/issues/331