chore(catalog): Load both full and filtered Camel Catalog
Currently, we're loading the short version of the Camel Models Catalog
to not only populate the UI catalog but also as a lookup registry when
building the Configuration Form.

This is not ideal since while having less elements in the catalog it also
means less available definitions, render some processors not configurable.

This commit loads both catalogs and use the short list to build the UI Catalog
and the full one for the lookup.