fix(canvas): Replace nodes that contains a single steps property don't work
Currently, there's a piece of logic that checks if the node meant to be replaced has a single
`step` property that can have children. This logic was causing that for some special nodes,
like `doCatch`, the replace step functionality didn't work, and the intended replacement
step gets appended instead.

The fix is to remove said functionality as it was already implemented in a different manner.

Fixes: https://github.com/KaotoIO/kaoto-next/issues/370