fix(canvas): Avoid rerendering the entire application when configuring nodes
Currently, when configuring a node in the canvas, it refresh the source code. This in turn
force rerendering the entire application, as the SourceCode provider holds the source code
state and its placed at the root.

The fix is to split the provider into SourceCodeContext and SourceCodeApiContext, this way
the consumers can subscribe only to the required section avoiding having a reference to the
changing source code.

A few consumers examples:
* FlowClipboard component: it only needs access to the source code.
* NewFlow: it only needs access to the SourceCodeApi to signal that something changed
* Shell: it only needs access to the `entities:updated` signal to save the code to localStorage
* KaotoEditor and SourceCodePage: needs both context

Fixes: https://github.com/KaotoIO/kaoto-next/issues/213