Fix Catalog tag selection in VS Code
In VS Code, the onClick event was not provided to the Label of the
CatalogTagsPanel but transferred to the Tile of the Catalog. We have not
found the exact reason but providing the onClick on the `a` element used
to render the React component is working in both environments, browsers
and VS Code.

There are already UI tests in catalog.cy.ts -> `Catalog filtering using
tags`

fixes #431

Signed-off-by: Aurélien Pupier <apupier@redhat.com>