fix(ExpressionEditor): Fix stale data in Expression Editor
Currently, the ExpressionEditor and DataFormatEditor hold stale data
about the selected node, hence forcing the node update with old data.

The issue behaves as the following:

1. The user updates the expression field using the ExpressionEditor
2. After that, a change in an endpoint parameter is made
3. Another update is made in the ExpressionEditor
At this point, the ExpressionEditor doesn't know the latest change from
the endpoint parameters, hence updating the model with stale information
and ignoring the pending changes.

The fix is to refresh the data in the ExpressionEditor before making the
changes.

This issue will behave the same for the `DataFormatEditor`

fix: https://github.com/KaotoIO/kaoto-next/issues/518