fix(vscode): Fix VSCode switching integration type
Currently, when VSCode switch from one integration type to another, for
instance, going from a Camel Route to a Kamelet, this change is not
persisted until a property is changed.

This happens because the KaotoBridge is not subscribed to the
`code:updated` event and this is used when changing from one integration
type to another.

The fix is to subscribe to that envent. In addition to that, since [this
PR](https://github.com/KaotoIO/kaoto-next/pull/541), the
`VisibleFlowsProvider` was removed from the `Visualization` component,
hence it was breaking the VSCode integration. This PR also adds it to
the `KaotoBridge`

fixes: https://github.com/KaotoIO/kaoto-next/issues/402