feat(form): Parse URI field upon configuring step
Currently, we're not parsing the URI field from the `from`, `to`, and the `toD` Camel processors.

This pull request adds support for the URI field in the following fashion:
Considering the following `URI`: `timer:timer-1?period=5000&delay=5&synchronous=true`, when configuring the `timer` component, we parse the `URI` using the component's syntax (in this case is `timer:timerName`) and set those values in the parameter object. After the parsing, the `URI` field is cleaned to avoid overriding the parameter's value again. The result of this parsing is: `{ timerName: 'timer-1', period: 5000, delay: 5, synchronous: true }`.

In a more practical example, given this `yaml` route example:
```yaml
- from:
    uri: timer:timer-1?period=5000&delay=5&synchronous=true
    steps: []
```

after configuring the `timer` component, we'll end up with the following route:
```yaml
- route:
    from:
      uri: timer
      steps: []
      parameters:
        period: 5000
        delay: "50"
        synchronous: true
    id: route-3338
```

fix: https://github.com/KaotoIO/kaoto-next/issues/274