fix(schemas): Avoid mutating component schemas
When combining processor's schemas with the component ones, the former
is mutated to hold the later schema. This present an issue as the schema
gets modified everytime.

The fix is to clone the schema before combining it, so the original
reference is preserved.

fix: https://github.com/KaotoIO/kaoto-next/issues/631