Currently, we're removing parameters set by the user that match the default value intending to have less verbose routes.
There's an issue with the `file` component, detailed as follows:
```
from: file:someFile.txt
parameters:
  noop: true
```

In the previous route, `noop=true` implicitly set `idempotent=true` **IF THE VALUE HAS NOT BEEN SET TO EITHER TRUE OR FALSE**, usually this is OK and doesn't represent a problem.

In case the user wants to set `noop=true` and `idempotent=false` to force the `file` component to re-read the files when the scheduler kicks in, like the following:

```
from: file:someFile.txt
parameters:
  noop: true
  idempotent: false
```

Then we will remove the `idempotent=false` as it is the default value for the `idempotent` parameter, effectively blocking the user from setting a different behavior for the `file` component.

This commit aims to keep values set by the user, even in case they are the default ones.

fixes: https://github.com/KaotoIO/kaoto-next/issues/633