feat(sourcecode): Sort Camel Routes parameters
Currently, the order assigned to the YAML output is determined through the order of the properties' definition, meaning that for an object with a single `id` property, after creating another property, the order will be:

```yaml
id: something
anotherProperty: something else
```

Ideally, we should follow a consistent order to simplify the diffing between the route's versions.

This commit adds an intermediate step that sorts the parameters in the following order:

```
 *   1. `id` prop
 *   2. `description` prop
 *   3. `uri` prop
 *   4. `parameters` prop
 *   5. `steps` prop
 *   6. the remaining properties in alphabetical order
```

Later on, we might want to sort the `parameters` object using the order defined in the Camel Catalog.

fix: https://github.com/KaotoIO/kaoto-next/issues/639