fix(kamelets): Edit Kamelets' nodes properties
Currently, Kamelets' nodes can't be edited using the Canvas Form,
causing the changes made by the user, are not serialized into the
source code.

This bug was caused due to a problem in the serialization of the
`KameletResource`, because despite the `Canvas Form` editing the `KameletVisualEntity`
correctly, what ultimately was serialized was the original `Kamelet`
without changes.

The fix is to ensure that we're serializing the
`KameletVisualEntity.from` as part of the resulting `KameletResource.toJSON()` method.

This change opens up a more broad topic about `Kamelets`, at this moment,
the `KameletVisualEntity` class knows about the underlying flow
(`kamelet.spec.template.from`) and its metadata (`kamelet.metadata`), and
this shouldn't be the case.

Ideally, `KameletVisualEntity` would have knowledge about the flow only
but the issue with this is that it will close the possibility of
renaming the flow, as the `setId` method is defined in the
`KameletVisualEntity` class. A solution for this could be to create
an intermediate `setId` functionality in the `CamelResource` interface itself
which will later on delegate the change to the underlying
`VisualEntities.setId()` method if necessary, this way, in case of
`Kamelets` and `Pipes`, we have the chance to update its metadata and
also the underlying flow.

In addition to that, it would be convenient to offer a mechanism to sort
the properties of the underlying `VisualEntities`, this way we could
sort the entire `Kamelet` in a `kubernetes` friendly way and also
keeping the underlying `Camel Route` sorted differently.

fix: https://github.com/KaotoIO/kaoto-next/issues/594