feat(canvas): Show placeholder when there are no visible flows
Currently, when the canvas is empty, being because there are no flows
defined or because there are but hidden, we see a gray area. This might
present an issue for the users, starting with a blank file in VSCode, as
it's not clear what's the first step to start working with the app.

This commit introduces a placeholder in case there are no flows or there
are but not visible

fix: https://github.com/KaotoIO/kaoto-next/issues/362