feat(canvas): Validate nodes upon rendering
Currently, we're validating the canvas' nodes when editing the Form,
using the original model. This casuses that in some cases, we get a
false positive result since a parameter might have been defined using
the URI field.

This commit aims to validate each node upon rendering, meaning that the
node will be validated whenever the node is rendered, which happens the
first time when a route is presented or when the node is selected or
deselected.

In addition to that, we're parsing the node definition so we could
account for the URI field when applicable.

fix: https://github.com/KaotoIO/kaoto-next/issues/697