fix(setHeaders): Fix broken setHeaders processor
After changing the way how `AutoForm` updates the node's models, nested
schemas, like the one used by `setHeader` no longer work as the key of
the property being changed is expressed using the `.` (dot) notation.

As an example: `headers.0.expression`

This means that we can't use the `key` as it is to set the values, as it
won't achieve the expected result.

The fix is to use `lodash/set` to navigate through the path and assign
the value in the right place.

fix: https://github.com/KaotoIO/kaoto-next/issues/727