chore(deps): Update @patternfly/* dependencies
Updates:
*    `"@patternfly/patternfly": "5.2.0",`
*    `"@patternfly/react-code-editor": "5.1.0",`
*    `"@patternfly/react-core": "5.2.0",`
*    `"@patternfly/react-icons": "5.2.0",`
*    `"@patternfly/react-table": "5.2.0",`
*    `"@patternfly/react-topology": "5.2.0",`

Notes:
Notice that `@patternfly/react-code-editor` stayed at `5.1.0` since
there's an ongoing issue breaking the `YAML` autocompletion support.

The `@patternfly/react-topology` is being updated to `5.2.0` since is
required to draw squared containers in the canvas. Unfortunately,
there's an issue with a missing `pf-v5-u-screen-reader` class definition
which forces the Canvas' buttons to show their label.

In addition to that, `@patternfly/react-topology` changed the structure
of the nodes and the e2e needed to be updated. Now, once a node is
selected, the label gets promoted to a different layer, meaning that
there's no easy way to target the three dots menu to open the contextual
menu. In this case, this functionality was switched to a right click to
display the same contextual menu.

Relates to: https://github.com/KaotoIO/kaoto-next/issues/767