fix(ajv): Avoid ajv throw exception for unresolved fields
Currently, if a schema field cannot be resolved by using the `$ref`
keyword, `ajv` throws an exception, making the `CanvasForm` unusable.

In addition to that, the `form` field has been removed from the
`CanvasForm` as is meant to be rendered by the `Canvas` itself.

relates to: https://github.com/KaotoIO/kaoto-next/issues/767