fix(canvas-form): Avoid adding empty objects
Currently, when there's an object property in a schema, opening the
form, it gets the default value of `{}` (empty object) and is persisted to the source
code, even if the user didn't actually changed anything in the form.

The fix is to ignore such values when updating the model.

fix: https://github.com/KaotoIO/kaoto-next/issues/806