feat(canvas): Support root OnException
This PR adds support for the root `OnException` entity. This entity has
a root configuration that is linked through the `CanvasForm` component
but also has a `steps` property, effectively turning it into a
`VisualEntity`.

The `getVizNodeFromProcessor` method was extracted from the
`AbstractCamelVisualEntity` class in order to reuse it in the
`OnExceptionVisualEntity` class. In an upcoming PR, more methods will be
extracted in order to avoid duplication as much as possible.

There are more methods that could be extracted to avoid duplication, for instance:
1. `getNodeLabel()`
2. `getTooltipContent()`
3. `getComponentSchema()`
4. `updateModel()`

To make it possible, we would need to change how the `CamelRouteVisualEntity` is created,
as currently it requires a root `RouteDefinition` object, ideally, it should
receive a `{ route: RouteDefinition }` object instead.

fix: https://github.com/KaotoIO/kaoto-next/issues/561