feat(canvas): Add additional node's interactions
Currently, all canvas' nodes have the possibility to be replaced and
removed.

This leads to an inconsistent situation in which empty containers
suggests that can be replaced or removed.

The fix for this situation is to offer the possibility to select whether
a node can be replaced or removed, this way, we have a broad range of
node interactions.

fix: https://github.com/KaotoIO/kaoto-next/issues/561