chore(typings): Remove `ajv.JSONSchemaType` usage
Curently, we're relying on the `JSONSchemaType` from the `ajv` library,
and while this is working, we're not leveraging it in a meaningful way.

This commit removes its usage in favor of the lightweight `@types/json-schema` package
and consolidate its definition in the `KaotoSchema.ts` file, this way,
when a new type is desired, the change will happen in a single file
instead.