chore(form): Add SchemaBridge provider
Currently, we need to create a `JSONSchemaBridge` to use a schema with uniforms.

This commit creates a new `SchemaBridgeProvider` so the handling and creation of the bridge is consolidated in a single place, leaving the consumer to provide only a desired schema to the form.

relates: https://github.com/KaotoIO/kaoto-next/issues/560