fix(catalog): Provide fixed size for Properties Catalog
Currently, the Properties Modal size is determined according to its
content, meaning that long list will occupy the entire screen while
other tabs might be smaller, effectively changing the size and position
of the modal.

The fix for this is to provide a fixed size for the modal, according to
the User's viewport.

Ideally, in the future we should transform the properties modal into a
page, which will allow the user to share the link to a specific
component and properties. Ideally, it should also contain a breadcrumb
to show the user current location.

fix: https://github.com/KaotoIO/kaoto-next/issues/742