feat(catalog): Add tile counters
Currently, when looking for a `saga` component, we see a `Saga` component, but there's also a `Saga` pattern and there's no visual feedback that it exists, until we click on the `Pattern` toggle.

This commit adds counters to each toggle, so we can monitor the
occurrences in other groups as well.

In addition to that, the lookup is debounced until the user stop
writing, in a 500ms window. This helps when writing the first letters of
the search term, it doesn't get stuck until the UI renders.

The e2e tests for the catalog were updated to include a check after
resetting the search field.

fix: https://github.com/KaotoIO/kaoto-next/issues/745