feat(catalog): Sort Processor's properties
Currently, since the component's schema is generated from the Camel's components catalog,
iterating over the properties dictionary, the output is written using the same sorting as well.

Ths is not the case for processors, which are extracted from the Camel
YAML DSL schema first, and then decorated with elements from the Camel's
catalog.

This commit change the iteration to be performed from the Camel's
processor catalog instead.

Changes:
* Move the `class:xxx` information from `$comment` to `format`
* Move the `duration` information from `$comment` to `format`
* Add `format: password` for `secret: true` fields

fix: https://github.com/KaotoIO/kaoto-next/issues/51