fix(SourceCode): Avoid loading Schema multiple times
Currently, after loading the SourceCode editor, a specific schema is
applied to the monaco editor.

The problem with this is that going back and forth between the
SourceCode page and any other page, causes the schema to be loaded
multiple times, as the workers are not disposed.

The fix is to dispose the workers when leaving the SourceCode page.

fix: https://github.com/KaotoIO/kaoto-next/issues/833