fix(CanvasForm): Expression form hold stale data
When the users includes config in a form field, which has same name as form field of another
language/dataformat type, to which they can switch - the previously configured form field might get omited.

The fix is to assign a key to the form, so React know that the form
should be recreated without reusing the existing fields.

fix: https://github.com/KaotoIO/kaoto-next/issues/904