feat(canvasform): Use local SchemaBridge
Currently, in case a schema contains a `oneOf` array, `uniforms`
combines all `oneOf` definitions in a single schema definition.

The issue with this approach is that combine potentially non-compatible
schemas, like the `errorHandler` one, since we need to specify a single
property.

This commit extends the `getField` method from the uniforms `JSONSchemaBridge`
to add the `oneOf` definitions into the field, this way, we could use
this information in the form to render an UI control to select a given
schema

relates: https://github.com/KaotoIO/kaoto-next/issues/948
relates: https://github.com/KaotoIO/kaoto-next/issues/560