fix(code): Use YAML 1.1 for serializing
Currently, Kaoto serialize to YAML using YAML 1.2.

This creates an edge for the Rest component since one of its parameters
(`bindingMode`) accepts an `off` value which ultimately gets serialized as it is,
causing to be interpreted as `false`.

This commit change the YAML serializing schema to 1.1 so `off` gets serialized as `"off"`
avoiding this issue.

fix: https://github.com/KaotoIO/kaoto-next/issues/971