fix(Canvas): Restore contextual menu labels
After removing the node id from the contextual's menu labels, some EIPs have
duplicated `Insert` items, and that's not correct since one `Insert` is for
inserting additional nodes while the other is to insert special nodes, like `when` or `otherwise`.

This PR restores the previous label while we get a better approach from an UX perspective.