feat(Catalog): Add RouteConfiguration entities
As a prerequisite of https://github.com/KaotoIO/kaoto/issues/492, we
need to add the related entities to the catalog.

This commit adds the following entities:
* intercept
* interceptFrom
* interceptSendToEndpoint
* onException
* onCompletion

In addition to that, we also show the entity ID if provided in case
there's no description available.

relates: https://github.com/KaotoIO/kaoto/issues/492