feat(Canvas): Add new Entities to the canvas
Currently, setting the current DSL and creating a new entity are tied
together in a single step in the NewFlow component.

This commit split that functionality in two separate buttons:
1. DSL Selector: Meant to choose the current DSL and therefore determine
   which type of entities can be added into the Canvas
2. NewEntity: This is for adding new entities, supported by the current
   DSL, or said in a different way, allows to add entities from the
current CamelResource registry.

The existing functionality from the CanvasEmptyState is preserved for
now.

In addition to that, when removing all routes, a new route won't be
created by default.

fix: https://github.com/KaotoIO/kaoto/issues/1030