fix(Catalog): Using labels causes unintended navigation
Context:
Currently, we're using the `numLabels` property of the `LabelGroup` component to show just a few `Labels`.

When the user clicks on the `n more` label, a navigation is performed and the UI is reloaded, causing the VSCode extension to fail.

This happens because `LabelGroup` is a child of the `Form` component, causing to execute the default Form behavior when selecting a `Label`

Changes:
The fix is to move the Label bar outside of the `Form` component.

fix: https://github.com/KaotoIO/kaoto/issues/1073
fix: https://github.com/KaotoIO/kaoto/issues/1074