fix(CanvasForm): Provide basic schema for unknown Kamelets
Currently, when using the `kamelet` component, providing an unknown
Kamelet to Kaoto, this renders the basic `To` EIP configuration form.

This is not useful since then we can't provide the Kamelet's TemplateId
nor RouteId.

The fix is to fallback to the standard `Kamelet` component when passing
an unknown Kamelet to Kaoto.

fix: https://github.com/KaotoIO/kaoto/issues/1004