feat(Settings): Add Settings Page
In the scope of selecting the Runtime in use, a settings page pointing
to a catalog URL is needed.

This commit adds a Settings component and a Settings provider. By
default, it will store the settings to the browser's LocalStorage.

The idea behind using a SettingsAdapter is to give the possibilit in the
future to set the settings at VSCode level if needed.

fix: https://github.com/KaotoIO/kaoto/issues/834