fix(catalog-generator): Load the correct Catalog
Currently, the loaded Camel Catalog is the specified in the `pom.xml` file,
and this represents an issue when there are breaking changes introduced
like the one from `4.6.0` to `4.7.0` in which the `loadBalancers` were
renamed.

* [Camel 4.6.0 loadBalancers](https://github.com/apache/camel/blob/581b94a8d326e381d6c58cfe8ce80c92168ac29a/dsl/camel-yaml-dsl/camel-yaml-dsl/src/generated/resources/schema/camelYamlDsl.json#L2491-L2505)
* [Camel 4.7.0 loadBalancers](https://github.com/apache/camel/blob/a2212f39466d7ef9185b6557f872ca9a8a7da3b8/dsl/camel-yaml-dsl/camel-yaml-dsl/src/generated/resources/schema/camelYamlDsl.json#L2496-L2510)

This commit forces to load the correct Camel Catalog by means of
fetching the runtime Catalog with transitive dependencies, so the
appropriate classes are loaded as well.