fix(VSCode): Provide timeout and fallback for `KaotoEditorFactory`
Currently, VSCode Kaoto build with the main branch is broken due to a
missing API implementation that causes for a `promise` to hold forever,
preventing the extension to load.

The fix is to provide a `timeout` mechanism to unblock the promise, plus
a `fallback` mechanism to get sensitive settings default to continue the
extension bootstrap process.