feate(Canvas): Extend DagreGroupsLayout
Currently, `DagreLayout` is used to render the graph, and while it does
work, it doesn't support nested groups properly due to a limitation in
Dagre itself.

In the `@patternfly/react-topology` there's another layout called `DagreGroupsLayout` that performs individual layout inside each group, providing a better result when the graph contains nested groups.

This commit brings that class over and uses the initial node index for
sorting the graph node, this way the order is kept when expanding/collapsing nodes in the Canvas.

Related issue: https://github.com/patternfly/react-topology/issues/230