feat(Canvas): Add Node mapper
Currently, EIPs with children are rendered as containers, as opposed to in the past, where they were rendered as nodes.

The goal for this ticket is to be able to customize how each EIP is rendered.

This is an intermediate step between having a single class holding the
entire route vs having individual classes per EIP, each holding the
underlying code section they represent.

fix: https://github.com/KaotoIO/kaoto/issues/1329