chore(CanvasForm): Cleanup and File sorting
Currently we have a couple missing bits after implementing the "filter by
name" and "filter by category" functionality.

This commit simplifies the usage of `FormTabsModes` by removing both `enum`s and relying only in it's type.

It also moves the filters to the `CanvasFormHeader` and rename the
`CanvasFormTabs` component to `CanvasFormBody` to better convey its
purpose.