chore(VSCode): Refactor KaotoEditor structure
Currently, the `DesignPage` component is duplicated to serve the
multiplying architecture and the webapp.

In addition to that, for the multiplying architecture, we're using a
React render approach to navigate between different tabs, whereas in the
webapp we use a Router.

This commit removes the `Tabs` folder and use the `DesignPage` component directly to reduce the duplication.

In addition to that, a Router has been introduced to navigate between
the tabs, this as a cleanup for the upcoming tasks.