fix(MultiplyingArchitecture): Avoid setting stale edits
Currently, an MA edit lifecycle looks like the following:

1. The Kaoto editor makes an edit and notifies the MA.
2. After a quiet period (debounce), MA process the edit and sends it
   back to the Kaoto editor.
3. If the content is the same, this last call is ignored by Kaoto, as
   the editor content is exactly the same, so there's no need to
recreate the UI. If the contrary, the current state gets discarded and
the UI is regenerated with the incoming content.

This process opens the case for an edit happening a few milliseconds
before the MA sending the content back, meaning that at this stage, the
Kaoto content has changed but is receiving a stale content, causing for
the UI to be regenerated with the stale content.

The fix for this situation is to implement a hash for each content and
validate whether the incoming content is stale or not.

fix: https://github.com/KaotoIO/kaoto/issues/1376
fix: https://issues.redhat.com/projects/KTO/issues/KTO-452