fix(Canvas): Remove Autozoom functionality
Currently, when selecting a node in the canvas, there is an existing fuctionality to center the view.
This helps to keep the flow visible at all times, especially in cases when the side panel might overlap the flow.

With complex flows, this behavior turns to be problematic, as the user
might be zooming over a section of the flow, then selecting a node,
effectively moving the viewport somewhere else.

The fix is to remove the auto-center functionality so the zoom is ketp
between node editing.

fix: https://issues.redhat.com/browse/KTO-461