fix(Canvas): Fix Canvas nodes distorted after navigation
Currently, when using chromium-based browsers, navigating to from a page
to the Design page, makes the nodes render distorted.

There are a few elements contributing to this behavior, first of all,
the `visible-flow` provider was initialized to `{}`, causing to the `Canvas`
component to be rendered twice, and after that, it seems like the
`useEffect` callback that called `controller.fromModal` was batched
after the `VisualizationSurface` was initialized, missing the update.

The fix is to replace the `useEffect` callback with `useLayoutEffect` to
ensure the nodes and edges are available before the `VisualizationSurface`
component.

fix: https://github.com/KaotoIO/kaoto/issues/1352