chore(catalog-generator): Lower the required Java version to 17
Currently, the `catalog-generator` requires Java 21.

While that enable modern syntax, not all users have it installed in
their machines, so in order to lower the contribution requirements,
we're lowering the required version to be Java 17.

By default, in RHEL 9.0, Java 17 is preinstalled.

Here is an extract of the available Java version per RHEL version

| RHEL Version | Java version | Release date |
| --- | --- | --- |
| RHEL 7 | Java 8  | June 10, 2014 |
| RHEL 8 | Java 11 | May 7, 2019   |
| RHEL 9 | Java 17 | May 17, 2022  |

fix: https://github.com/KaotoIO/kaoto/issues/1476