Reduce Maven download log
On CI, the log cannot be correctly displayed and parsed in the GitHub
default UI due to the large amount of it, mainly due to the download
transfer progress of Maven components. Consequently, using
`--no-transfer-progress` on the first Maven build to reduce the amount
of logs.

Also updated Maven lifecycle phase, removed the verify phase as it is
included in the install phase. See
https://maven.apache.org/guides/introduction/introduction-to-the-lifecycle.html#a-build-lifecycle-is-made-up-of-phases
for reference

Signed-off-by: Aurélien Pupier <apupier@redhat.com>