fix(Canvas): Fix Edge source misplaced for containers
In order to make room for the node's secondary label, the containers
were expanded by `10px`.

The problem comes when in vertical layout, that additional size is not
taken into account when calculating the edges since the size is only
reflected in the visualization part and not during the layout
calculation.

The solution for now is to increase the group padding just a bit to make
the secondary label to fit.

fix: https://github.com/KaotoIO/kaoto/issues/1575