fix(DataMapper): Fix XPath Parser to tokenize the word starting with the reserved keyword char sequence, but longer
Fixes: https://github.com/KaotoIO/kaoto-datamapper-integration/issues/65
Fixes: https://github.com/KaotoIO/kaoto-datamapper-integration/issues/132
Fixes: https://github.com/KaotoIO/kaoto-datamapper-integration/issues/139

This example demonstrates the exact solution
https://github.com/Chevrotain/chevrotain/blob/master/examples/lexer/keywords_vs_identifiers/keywords_vs_identifiers.js

There is `longer_alt` in lexer option to mark the token to fallback to the other token if it's a longer one. For example, while there is `or` keyword, if there is `order` as a field name, we don't expect it to be tokenized as `or` and `der`. In order to do this fallback, `longer_alt` has to be specified for each shorter reserved words.