fix(Catalog): Ignore missing `model` properties
In the following Camel versions, a new `errorHandler` property was added
to the `RouteDefinition` to allow different `errorHandlers` per route.

* 4.4.x
* 4.5.0

The issue with that is the `errorHandler` property is present in the
YAML schema but not in the `RouteDefinition` model, making the
generation fail.

As a workaround, this commit ignores the missing properties from the model, but present in the YAML DSL.

The result of this is the `errorHandler` property missing in the
configuration form but present in the YAML DSL.