feat(DataMapper): Use Browser tooltips
Currently, we're using patternfly `<Tooltip>` component, and while it
works great for simple cases, in case there is a close component to the
one that is triggering the tooltip, the former gets hidden behind it,
making it complicated to interact with it.

The alternative is to use the Browser tooltip, leveraging the [title HTML attribute](https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/title)
if available.

fix: https://github.com/KaotoIO/kaoto-datamapper-integration/issues/184