fix(DataMapper): Avoid same navigation
Currently, the DataMapper tab shows a tutorial explaining how to use the
DataMapper.

The issue with that is that when the user is already in the DataMapper
editor, that tab navigates to the tutorial again.

The fix is to stop the same navigation in case the user is already in
the DataMapper tab.