fix(Canvas): Increase hoverOut timing
Currently, when opening the Step toolbar, the hover area is the same
size as the node size and this is problematic for steps that generates a
wider toolbar than the node itself, as in order to keep the toolbar
open, the user need to follow a `L` shaped path.

The fix is to increase the hoverOut delay time upon hovering the node, so the
user have more time to get to the rightmost or leftmost
button in the toolbar.

fix: https://github.com/KaotoIO/kaoto/issues/1631